pyproject.toml
[project]
name = "dsv"
version = "0.1.0"
description = "RDF Data-Shape-View System"
authors = [{ name = "Lucy Patton", email = "lucy.patton.368@gmail.com" }]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "rdflib (>=7.1.4,<8.0.0)",
  "pydantic (>=2.11.7,<3.0.0)",
  "fastapi (>=0.116.1,<0.117.0)",
  "uvicorn (>=0.35.0,<0.36.0)",
  "click (>=8.2.1,<9.0.0)",
  "watchdog (>=6.0.0,<7.0.0)",
  "rich (>=14.1.0,<15.0.0)",
  "pyshex (>=0.8.1,<0.9.0)",
  "rdflib-jsonld (>=0.6.2,<0.7.0)",
  "oxigraph (>=0.4.11,<0.5.0)",
  "sparqlwrapper (>=2.0.0,<3.0.0)",
  "docker (>=7.1.0,<8.0.0)",
  "requests (>=2.32.5,<3.0.0)",
  "python-dotenv (>=1.1.1,<2.0.0)",
]

[tool.poetry]

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
black = "^25.1.0"
mypy = "^1.17.1"
flake8 = "^7.3.0"

types-requests = "^2.32.4.20250809"
types-docker = "^7.1.0.20250822"

[tool.black]
line-length = 100
target-version = ["py310"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | build
  | dist
)/
'''

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
